Fixed it so that the RabbitMQ process will daemonize properly
You'll see setsid used in all the init statements in this cookbook.
This is because there is a problem with the stock init script in the
RabbitMQ debian package (at least in 2.8.2) that makes it not daemonize
properly when called from chef. The setsid command forces the
subprocess into a state where it can daemonize properly. (thanks to
Daniel DeLeo for the help)