Use the Redhat service provider.
The recipe didn't work with the  RabbitMQ distro package in Fedora 19 because it doesn't have an init.d script anymore. They're using systemd/systemctl now, so I switched to the Redhat provider for the service resource. That still uses the old 'service ' command, but seems the most compatible option for now.