enable pause_minority mode in rabbitmq config file
per http://www.rabbitmq.com/partitions.html, if you add
cluster_partition_handling=pause_minority
to the config file, the rabbit cluster will, when it loses a member,
have that member pause itself until it can see the others again, then
restart so it can resume message passing as normal. We needed this
because the nodes that were disconnecting (causing "network
partitioning") were staying in their sad, defunct state indefinitely and
awaiting user intervention in the form of a reboot (which the
instructions say is the correct way to heal a network partition anyway).
So this just goes ahead and does that without waiting for some big
important human to give his blessing or whatever.  From what I've
gathered, the default setting for this variable is ignore, so leaving
that as default here and getting the value we actually want via
data_bags from chef-repo/data_bags/platform/rabbitmq/config_demo.json