Refactored the template action when setting the erlang cookie for a clustered node to ensure the service is restarted before attempting to stop and start the running application.
It is hard to debug but it would appear that the "notifies" method appears to run out of order and attempts to run the execute 'reset-node' before the service is fully stopped and restarted. Manually attempting to reset the node now fails as the running services cookies is different from the configured one.