Chefspec is the best spec
This is an inital commit of some generic chefspec
tests for the rabbitmq cookbook. I've tied it up to
the `rake travis` task so when this is merged we can
have it run as part of the PR process.

Again this is just an example and a start, we can attempt
to test the LWRPs at a later date.

Comments are more than welcome.